Regroup shared parts between FlowBlocks
